Up to date
I modified the plc converter algorithm. Now it is compatible with
real phisics limit switches. It works on CODESYS:
the limit switches are like manual switches in CODESYS, it means
that the code will stop on that IF statement until the switch is
triggered in real life.

I let the old plc.bak.py file just for tests.