Up to date
New version of the plc converter alghorithm. Now it works as the plc reads the program in loop from start to finish, until the IF statements are false.
I made the IF statements so that the inputs and outputs are not accessible until the Real Life limit switches are triggered.
This prevents the errors when multiple signals arrive to the same electrovalve.