PLC ST code converter
Up to date
Now it is possible to connect through OPC DA Server from CodeSyS to FluidSim.
With my ST code you will be able to control the FluidSim PLCs from the CODESYS Software with a local OPC DA Server.
I will upload a tutorial pdf to install all the dependencies and programs to make it work.
Enjoy it.